move code from trainer to tuner
